Moved renderers back to the rendering package.
